python-validator/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.0.0
click>=8.1.0
ipaddress
